Added tags in post creation + moved the notification subscriber to navbar
